backend/requirements.txt
fastapi[standard]
sqlalchemy
pydantic
passlib[bcrypt]
pyjwt